fix balance for cumulative savings
Balance buttons will now balance for savings value, not for starting amount, so the user can balance for example monthly savings repeatedly during the sim to maintain purchasing power. Otherwise due to inflation the purchasing power will still slowly start to leak away as the monthly amounts are not enough anymore to compensate.